Python-DAGs/Hello-World/requirements.txt
# requests==2.32.3
# graphviz==0.20.3
apache-airflow==2.10.4
psycopg2-binary==2.9.11
SQLAlchemy==1.4.54
sqlalchemy-utils==0.42.0
alembic==1.17.0
pytest==8.4.2
flake8==7.3.0
load_dotenv==0.1.0